Trait system_interface::io::IsTerminal

pub trait IsTerminal {
    fn is_terminal(&self) -> bool;
}

Extension for I/O handles which may or may not be terminals.

Required methods

fn is_terminal(&self) -> bool

Test whether this output stream is attached to a terminal.

This operation is also known as isatty.