Trait system_interface::io::Peek

pub trait Peek {
    fn peek(&mut self, buf: &mut [u8]) -> Result<usize>;
}

A trait providing the peek function for reading without consuming.

Many common Read implementations have Peek implementations, however Stdin, ChildStderr, ChildStdout, PipeReader, and CharDevice do not, since they are unbuffered and pipes and character devices don’t support any form of peeking.

Required methods

fn peek(&mut self, buf: &mut [u8]) -> Result<usize>

Reads data from a stream without consuming it; subsequent reads will re-read the data. May return fewer bytes than requested; Ok(0) indicates that seeking is not possible (but reading may still be).