Trait system_interface::io::ReadReady

pub trait ReadReady {
    fn num_ready_bytes(&self) -> Result<u64>;
}

Extension for readable streams that can indicate the number of bytes ready to be read immediately.

Required methods

fn num_ready_bytes(&self) -> Result<u64>

Return the number of bytes which are ready to be read immediately.